+++ src/angelica/solvers/__init__.py
@@ -1,6 +1,7 @@
 from .base import BaseSolver
 from .steady_black_oil import BlackOilSolverSettings, SteadyBlackOilSolver
 from .steady_compressible import CompressibleSolverSettings, SteadyCompressibleSolver
+from .steady_compositional import CompositionalSolverSettings, SteadyCompositionalSolver
 from .steady_isothermal_incompressible import SteadyIsothermalIncompressibleSolver
 from .steady_non_isothermal_incompressible import (
     NonIsothermalSolverSettings,
@@ -10,9 +11,11 @@ from .steady_non_isothermal_incompressible import (
 __all__ = [
     "BaseSolver",
     "BlackOilSolverSettings",
+    "CompositionalSolverSettings",
     "CompressibleSolverSettings",
     "NonIsothermalSolverSettings",
     "SteadyBlackOilSolver",
+    "SteadyCompositionalSolver",
     "SteadyCompressibleSolver",
     "SteadyIsothermalIncompressibleSolver",
     "SteadyNonIsothermalIncompressibleSolver",